remove impedance 3D as the 6D should take care of it
