+++ demos/demo_impedance_ctrl_cpp.py
@@ -1,20 +1,18 @@
-##################################################################################################################
-## This file is the centroidal controller demo for solo12 using the C++ code.
-#################################################################################################################
-## Author: Julian
-## Date: 01/21/2021
-#################################################################################################################
+"""demo_impedance_ctrl_cpp
+
+This file is the impedance controller demo for solo12 using the C++ code.
+
+Author: Julian
+Date: 01/21/2021
+"""
 
 import numpy as np
-import time
-import os
-
 np.set_printoptions(precision=2, suppress=True)
 
-import pinocchio as pin
+import pinocchio
 from bullet_utils.env import BulletEnvWithGround
 from robot_properties_solo.solo12wrapper import Solo12Robot, Solo12Config
-import mim_control_cpp
+from mim_control_cpp import ImpedanceController
 
 
 if __name__ == "__main__":
@@ -31,7 +29,7 @@ if __name__ == "__main__":
     # ###################### impedance controller demo #################################
     root_names = ["base_link", "base_link", "base_link", "base_link"]
     endeff_names = ["FL_FOOT", "FR_FOOT", "HL_FOOT", "HR_FOOT"]
-    ctrls = [mim_control_cpp.ImpedanceController3D() for _ in endeff_names]
+    ctrls = [ImpedanceController() for _ in endeff_names]
     for i, ctrl in enumerate(ctrls):
         ctrl.initialize(pin_robot.model, root_names[i], endeff_names[i])
 
@@ -50,10 +48,10 @@ if __name__ == "__main__":
         -0.147,
         -0.2,
     ]
-    xd_des = pin.Motion.Zero()
-    kp = [200, 200, 200]
-    kd = [10.0, 10.0, 10.0]
-    f = pin.Force.Zero()
+    xd_des = pinocchio.Motion.Zero()
+    kp = np.array([200, 200, 200, 0, 0, 0])
+    kd = np.array([10.0, 10.0, 10.0, 0, 0, 0])
+    f = pinocchio.Force.Zero()
 
     end_eff_ids = [
         pin_robot.model.getFrameId(endeff_name) for endeff_name in endeff_names
@@ -68,8 +66,8 @@ if __name__ == "__main__":
         q0[1] += 0.3
         q0[1] += 0.1
         q0[2] += 0.2
-        q0[3:7] = pin.Quaternion(
-            pin.rpy.rpyToMatrix(
+        q0[3:7] = pinocchio.Quaternion(
+            pinocchio.rpy.rpyToMatrix(
                 np.array([float(n) / float(N) * 2 * np.pi, 0, 0])
             )
         ).coeffs()
@@ -85,7 +83,7 @@ if __name__ == "__main__":
                     kp,
                     kd,
                     0.0,
-                    pin.SE3(np.eye(3), np.array(x_des[3 * i : 3 * (i + 1)])),
+                    pinocchio.SE3(np.eye(3), np.array(x_des[3 * i : 3 * (i + 1)])),
                     xd_des,
                     f,
                 )
