+++ srcpy/mim_control.cpp
@@ -11,8 +11,7 @@
 
 namespace mim_control
 {
-void bind_impedance_controller6d(pybind11::module &module);
-void bind_impedance_controller3d(pybind11::module &module);
+void bind_impedance_controller(pybind11::module &module);
 void bind_centroidal_pd_controller(pybind11::module &module);
 void bind_centroidal_force_qp_controller(pybind11::module &module);
 
@@ -28,8 +27,7 @@ PYBIND11_MODULE(mim_control_cpp, m)
            subtract
     )pbdoc";
 
-    bind_impedance_controller3d(m);
-    bind_impedance_controller6d(m);
+    bind_impedance_controller(m);
     bind_centroidal_pd_controller(m);
     bind_centroidal_force_qp_controller(m);
 }
